feat: improved documentation
