+++ src/main/java/de/gematik/service/SubscriptionHandshakeFinalizer.java
@@ -8,16 +8,30 @@ import lombok.extern.slf4j.Slf4j;
 import org.hl7.fhir.r4.model.*;
 import org.springframework.stereotype.Service;
 
+/**
+ * Service responsible for finalizing the status of a FHIR Subscription resource
+ * after a handshake process. This is executed in a new transaction, typically
+ * off the request thread, to ensure the final status is committed independently.
+ */
 @Slf4j
 @Service
 @RequiredArgsConstructor
 public class SubscriptionHandshakeFinalizer {
 
+    /**
+     * Registry for accessing FHIR resource DAOs.
+     */
 	private final DaoRegistry daoRegistry;
 
   /**
-   * Commit the final status in an independent transaction.
-   * Runs OFF the request thread (caller executes via an Executor).
+     * Finalizes the status of a Subscription resource by updating its status and removing
+     * a specific tag, all within a new transaction. This method is intended to be called
+     * asynchronously (e.g., via an Executor).
+     *
+     * @param subscriptionId the ID of the Subscription resource to update
+     * @param markSys the system of the tag to remove
+     * @param markCode the code of the tag to remove
+     * @param ok whether the handshake was successful (true for ACTIVE, false for ERROR)
    */
   @Transactional(Transactional.TxType.REQUIRES_NEW)
 	public void finalizeStatus(String subscriptionId, String markSys, String markCode, boolean ok) {
