+++ src/main/java/de/gematik/service/SubscriptionHeartbeatService.java
@@ -21,25 +21,46 @@ import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.stream.Collectors;
 
+/**
+ * Service responsible for periodically sending heartbeat notifications for active FHIR Subscriptions.
+ * Determines which subscriptions are due for a heartbeat based on their configured interval,
+ * dispatches heartbeats per topic, and maintains the last sent time for each subscription.
+ */
 @Slf4j
 @Service
 @RequiredArgsConstructor
 public class SubscriptionHeartbeatService {
 
+    /**
+     * URL of the backport heartbeat period extension.
+     */
 	private static final String EXT_HEARTBEAT =
 		"http://hl7.org/fhir/uv/subscriptions-backport/StructureDefinition/backport-heartbeat-period";
 
+    /**
+     * Registry for accessing FHIR resource DAOs.
+     */
 	private final DaoRegistry daoRegistry;
-	private final TopicNotifyService topicNotifyService;
 
-	// Letztes Heartbeat je Subscription
+    /**
+     * Service for dispatching heartbeat notifications.
+     */
+	private final HeartBeatDispatchService topicNotifyService;
+
+    /**
+     * Stores the last heartbeat sent time for each subscription (by ID).
+     */
 	private final Map<String, Instant> lastSent = new ConcurrentHashMap<>();
 
-	/** Alle 60s prüfen; tatsächlicher Versand richtet sich nach dem per-Subscription-Intervall. */
+    /**
+     * Periodically checks all active subscriptions and sends heartbeat notifications
+     * for those that are due, based on their configured heartbeat period.
+     * Runs every 60 seconds; actual dispatch depends on each subscription's interval.
+     */
 	@Scheduled(fixedDelayString = "PT60S")
 	@Transactional
 	public void run() {
-		// 1) Aktive Subscriptions laden
+        // 1) Load active subscriptions
 		IFhirResourceDao<Subscription> subDao = daoRegistry.getResourceDao(Subscription.class);
 		SystemRequestDetails srd = new SystemRequestDetails();
 
@@ -49,7 +70,7 @@ public class SubscriptionHeartbeatService {
 		List<IBaseResource> resources =
 			subDao.search(map, srd).getAllResources();
 
-		// 2) Aus Subscriptions: Topic + Heartbeat-Periode lesen, „fällige“ Subscriptions pro Topic sammeln
+        // 2) For each subscription, read topic and heartbeat period, collect due subscriptions per topic
 		Map<String, List<Subscription>> dueByTopic = resources.stream()
 			.map(r -> (Subscription) r)
 			.map(sub -> new SubWithMeta(sub,
@@ -60,7 +81,7 @@ public class SubscriptionHeartbeatService {
 			.collect(Collectors.groupingBy(SubWithMeta::topic,
 				Collectors.mapping(SubWithMeta::sub, Collectors.toList())));
 
-		// 3) Pro Topic genau ein Dispatch mit leerer Ressourcenliste (keine Filterkosten)
+        // 3) For each topic, dispatch a single heartbeat (empty resource list)
 		Instant now = Instant.now();
 		for (Map.Entry<String, List<Subscription>> entry : dueByTopic.entrySet()) {
 			String topic = entry.getKey();
@@ -69,7 +90,7 @@ public class SubscriptionHeartbeatService {
 
 			int queued = topicNotifyService.dispatchHeartbeat(topic);
 
-			// 4) Nur wenn etwas gequeued wurde, „lastSent“ für die fälligen Subscriptions aktualisieren
+            // 4) If a heartbeat was dispatched, update lastSent for the due subscriptions
 			if (queued > 0) {
 				for (Subscription s : dueSubs) {
 					String key = s.getIdElement().toUnqualifiedVersionless().getValue();
@@ -78,18 +99,29 @@ public class SubscriptionHeartbeatService {
 			}
 		}
 
-		// 5) Aufräumen: Einträge entfernen, deren Subscriptions nicht mehr aktiv sind
+        // 5) Cleanup: remove entries for subscriptions that are no longer active
 		cleanupStaleEntries(resources);
 	}
 
+    /**
+     * Determines if a subscription is due for a heartbeat based on its last sent time and period.
+     *
+     * @param m subscription metadata
+     * @return true if the subscription is due for a heartbeat, false otherwise
+     */
 	private boolean isDue(SubWithMeta m) {
 		String key = m.sub().getIdElement().toUnqualifiedVersionless().getValue();
 		Instant last = lastSent.getOrDefault(key, Instant.EPOCH);
 		long elapsed = Duration.between(last, Instant.now()).getSeconds();
-		// Mini-„grace“ von 2s gegen Scheduler-Drift
+        // Add a small grace period (2s) to account for scheduler drift
 		return elapsed + 2 >= m.periodSeconds();
 	}
 
+    /**
+     * Removes entries from lastSent for subscriptions that are no longer active.
+     *
+     * @param activeSubs list of currently active subscriptions
+     */
 	private void cleanupStaleEntries(List<IBaseResource> activeSubs) {
 		Set<String> activeIds = activeSubs.stream()
 			.map(r -> r.getIdElement().toUnqualifiedVersionless().getValue())
@@ -97,22 +129,29 @@ public class SubscriptionHeartbeatService {
 		lastSent.keySet().removeIf(id -> !activeIds.contains(id));
 	}
 
+    /**
+     * Extracts the canonical topic URL from the subscription criteria if it is a valid URI.
+     * Returns null if the criteria is not a canonical URL.
+     *
+     * @param criteria the subscription criteria string
+     * @return the canonical topic URL, or null if not valid
+     */
 	private static String extractBackportCanonicalOrNull(String criteria) {
 		if (criteria == null) return null;
 		String c = criteria.trim();
 		if (c.isEmpty()) return null;
 
-		// harte Ausschlüsse: sieht wie eine klassische R4-Search-Query aus
+        // Exclude classic R4 search queries
 		if (c.contains("?") || c.contains("&") || c.contains("=") || c.contains(" ")) {
 			return null;
 		}
 
-		// muss http/https sein
+        // Must start with http/https
 		if (!(c.startsWith("http://") || c.startsWith("https://"))) {
 			return null;
 		}
 
-		// syntaktisch gültige URI?
+        // Check for syntactically valid URI
 		try {
 			new URI(c);
 		} catch (URISyntaxException e) {
@@ -122,7 +161,12 @@ public class SubscriptionHeartbeatService {
 		return c;
 	}
 
-	/** Heartbeat-Periode (Sekunden) aus der Backport-Extension am channel lesen. */
+    /**
+     * Reads the heartbeat period (in seconds) from the backport extension on the subscription channel.
+     *
+     * @param sub the Subscription resource
+     * @return the heartbeat period in seconds, or null if not present
+     */
 	private static Integer readHeartbeatPeriodSeconds(Subscription sub) {
 		if (sub.getChannel() == null) return null;
 		for (Extension ext : sub.getChannel().getExtension()) {
@@ -133,5 +177,12 @@ public class SubscriptionHeartbeatService {
 		return null;
 	}
 
+    /**
+     * Record holding a subscription and its associated topic and heartbeat period.
+     *
+     * @param sub the Subscription resource
+     * @param topic the canonical topic URL
+     * @param periodSeconds the heartbeat period in seconds
+     */
 	private record SubWithMeta(Subscription sub, String topic, Integer periodSeconds) {}
 }
